move long table parsing to utils
